requirements_build.txt
# Requirements for building executable package
pyinstaller==6.14.1
pillow==11.2.1
psutil==7.0.0
requests==2.32.4

# Core application dependencies
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
flask-login==0.6.3
werkzeug==3.0.3
email-validator==2.2.0
openpyxl==3.1.5
psycopg2-binary==2.9.9
gunicorn==23.0.0
wtforms==3.1.2
sqlalchemy==2.0.32